Remove `*.netstandard.cs` file suffix
The .NET MAUI Community Toolkit doesn't support .NET Standard; only .NET 6.0+ is supported.

We will defer this to the Discussion #301:
https://github.com/CommunityToolkit/Maui/discussions/301
